+++ src/components/@SignForm/PasswordStrength/index.ts
