+++ src/pages/@SignPages/SignInPage/SignInPage.tsx
@@ -1,0 +1,10 @@
+import { SignInForm } from '@/components/@SignForm/SignInForm'
+import SignGroupLayout from '../SignLayout'
+
+export default function SignInPage() {
+  return (
+    <SignGroupLayout>
+      <SignInForm />
+    </SignGroupLayout>
+  )
+}
