+++ src/pages/@SignPages/SignUpPage/SignUpPage.tsx
@@ -1,0 +1,10 @@
+import { SignUpForm } from '@/components/@SignForm/SignUpForm'
+import SignGroupLayout from '../SignLayout'
+
+export default function SignUpPage() {
+  return (
+    <SignGroupLayout>
+      <SignUpForm />
+    </SignGroupLayout>
+  )
+}
