+++ src/pages/@SignPages/index.ts
@@ -1,0 +1,2 @@
+export * from './SignInPage'
+export * from './SignUpPage'
