--- src/pages/SignUpPage/SignUpPage.tsx
@@ -1,16 +1,0 @@
-import { Footer } from '@/components/Footer/Footer'
-import { Header } from '@/components/Header'
-import { SignInForm } from '@/components/SignInForm'
-import { Container, Stack } from '@mui/material'
-
-export default function SignUpPage() {
-  return (
-    <Stack>
-      <Header />
-      <Container>
-        <SignInForm />
-      </Container>
-      <Footer />
-    </Stack>
-  )
-}
