+++ src/router/router.tsx
@@ -1,5 +1,7 @@
 import { ErrorBoundary } from '@/components/ErrorBoudary'
 import { RouterPath } from '@/constants/constants'
+import { SignInPage } from '@/pages/@SignPages/SignInPage'
+import { SignUpPage } from '@/pages/@SignPages/SignUpPage'
 import { ErrorPage } from '@/pages/ErrorPage'
 import { MainPage } from '@/pages/MainPage'
 import { NotFoundPage } from '@/pages/NotFoundPage'
@@ -35,6 +37,22 @@ export const router = createBrowserRouter(
         element={
           <ErrorBoundary fallback={<ErrorPage />}>
             <SignInUpPage />
+          </ErrorBoundary>
+        }
+      />
+      <Route
+        path={RouterPath.SignIn}
+        element={
+          <ErrorBoundary fallback={<ErrorPage />}>
+            <SignInPage />
+          </ErrorBoundary>
+        }
+      />
+      <Route
+        path={RouterPath.SignUp}
+        element={
+          <ErrorBoundary fallback={<ErrorPage />}>
+            <SignUpPage />
           </ErrorBoundary>
         }
       />
